refactor result checking
Results may not be posted by the time the bot detects an event has finished so results are now scheduled as a separate check process. The bot will not delete events from the database if their results have not yet been posted unless it has been over 2 months since the event ended, at which point it will assume the results won't be posted and remove the event anyway.